+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Add
+
+- Attraction component
+- Agenda component
+
 ## [0.0.71] - 2024-09-24
 
 ### Fix